feat(adaptive): add adaptive for mobile devices
consta-design-system/uikit#2349
