+++ src/containers/StandPage/StandPageFooter/StandPageFooter.css
@@ -27,11 +27,14 @@
     align-items: flex-start;
     width: 100%;
     margin-top: var(--space-s);
+    flex-wrap: wrap;
+    gap: var(--space-s);
   }
 
   &-Buttons {
-    & > *:not(:last-child) {
-      margin-right: var(--space-l);
-    }
+    display: flex;
+    justify-content: flex-end;
+    flex-wrap: wrap;
+    gap: var(--space-l);
   }
 }
